refactor(frontend): unify network topology types in shared/types/network.ts
